requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
gunicorn==23.0.0
whitenoise==6.8.2
python-dotenv==1.0.1
dj-database-url==1.2.0